+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 # Version 16
 
 ### [Current changes]
+- **[Add]** Added -spu or --stop_pip_auto_update to disable auto python module update
 - **[Add]** Immediate parent support in windows path
+- **[Add]** Add Open Empty Browser Action
 - **[Add]** Fix browserstack issue from node and refactored for appium v2
 
 ### [16.1.0][Oct 25, 2022]